# openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EvoNEST API",
    "version": "2.0.0",
    "description": "\n# EvoNEST General Documentation\n\n**The EvoNEST User Guide and the EvoNEST Developer Guide are found here**: https://daniele-liprandi.github.io/EvoNEST-backbone/\n\n# EvoNEST API Documentation\n\nThe EvoNEST API allows users to interact with the MongoDB databases generated in their NEST, plus it gives access to utility functions useful in ecological, evolutionary and biological research.\n\n## Features\n\n- **Samples Management**: Create, update, and retrieve biological samples with taxonomic information\n- **Traits Measurement**: Record and analyze trait measurements with statistical calculations\n- **Experiments**: Manage measurements, documents and raw data\n- **File Storage**: Upload and manage research files with metadata\n- **User Management**: Handle user accounts and permissions\n- **Utilities**: Geocoding, species image search, and data processing tools\n\n## Authentication\n\nMost endpoints require proper authentication and database access permissions.\n\n## Data Format\n\nAll endpoints return JSON data unless otherwise specified. Dates are in ISO 8601 format.\n      ",
    "contact": {
      "name": "Daniele Liprandi",
      "email": "daniele.liprandi@gmail.com",
      "url": "https://github.com/daniele-liprandi/EvoNEST-backbone"
    },
    "license": {
      "name": "GNU Affero General Public License v3.0",
...
    },
...
    "x-build-context": "GitHub Actions"
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionAuth": []
    }
  ],
  "paths": {
...
    "/api/samples/ext": {
      "get": {
        "summary": "Export samples data with API key authentication",
        "description": "Export all samples from a database using API key authentication. Supports various filtering and relationship inclusion options.",
        "tags": [
          "Samples"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "database",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Database name to export from"
          },
          {
            "in": "query",
            "name": "apiKey",
            "schema": {
              "type": "string"
            },
            "description": "API key for authentication (can also use Authorization header)"
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "default": "json"
            },
            "description": "Export format (only JSON supported)"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "animal",
                "silk",
                "other"
              ]
            },
            "description": "Filter samples by type"
          },
          {
            "in": "query",
            "name": "includeRelated",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include parent sample data chain"
          }
        ],
        "responses": {
          "200": {
            "description": "Samples exported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "No samples found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/experiments/ext": {
      "get": {
        "summary": "Export experiments data with API key authentication",
        "description": "Export all experiments from a database using API key authentication. Supports filtering and inclusion of raw data and related samples.",
        "tags": [
          "Experiments"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "database",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Database name to export from"
          },
          {
            "in": "query",
            "name": "apiKey",
            "schema": {
              "type": "string"
            },
            "description": "API key for authentication (can also use Authorization header)"
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "default": "json"
            },
            "description": "Export format (only JSON supported)"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filter experiments by type"
          },
          {
            "in": "query",
            "name": "includeRawData",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include raw experimental data"
          },
          {
            "in": "query",
            "name": "includeOriginalData",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include original unprocessed data"
          },
          {
            "in": "query",
            "name": "includeRelated",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include related sample data chain"
          }
        ],
        "responses": {
          "200": {
            "description": "Experiments exported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "No experiments found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}